- display now properly handled with floats to adjust for different resolutions - previewers are now the full size of 1 tile - board now begins at bottom of screen instead of with a y-offset - tidied up title screen - added new cool font!

+++ Assets/Puzlr/Scripts/TileDisplay.cs
@@ -62,7 +62,7 @@ public class TileDisplay : MonoBehaviour, IPuzlGameComponent
         }
     }
 
-    public void SetDisplaySize(int displaySize)
+    public void SetDisplaySize(float displaySize)
     {
         RectTransform thisRect = (RectTransform)transform;
         thisRect.sizeDelta = new Vector2(displaySize, displaySize);
